requirements.txt
pydantic-settings
loguru
sqlmodel
sqlalchemy
aiosqlite
greenlet

google-generativeai
pytest
pytest-asyncio
httpx
tenacity
pypdf
google-adk[eval]>=0.1.0
tenacity
python-telegram-bot[job-queue]>=20.0
python-dotenv

